Rename metadata.py -> _gjson.py
Looks like only geojson tools are left in that file now
